chore: rename rule.rs to rule_data.rs
For consistency
