+++ programs/transfer_limit/src/state/mod.rs
@@ -1,8 +1,8 @@
 mod config;
 mod member;
 mod record;
-mod rule;
+mod rule_data;
 
 pub use config::*;
 pub use member::*;
-pub use rule::*;
+pub use rule_data::*;
